+++ sbomify/apps/sboms/tests/test_vulnerability_scanning.py
@@ -10,7 +10,7 @@ from django.conf import settings
 from django.test import TestCase, override_settings
 
 from sbomify.apps.core.models import User
-from sbomify.tasks import scan_sbom_for_vulnerabilities_unified
+from sbomify.apps.vulnerability_scanning.tasks import scan_sbom_for_vulnerabilities_unified
 from sbomify.apps.sboms.models import SBOM, Component
 from sbomify.apps.teams.models import Team
 
@@ -37,7 +37,7 @@ class VulnerabilityScanningTimeoutTests(TestCase):
 
     @patch("sbomify.apps.core.object_store.S3Client")
     @patch("sbomify.apps.vulnerability_scanning.clients.subprocess.run")
-    @patch("sbomify.tasks.dramatiq.get_broker")
+    @patch("sbomify.apps.vulnerability_scanning.tasks.dramatiq.get_broker")
     def test_subprocess_timeout_handling(self, mock_broker, mock_subprocess, mock_s3_client):
         """Test that subprocess timeouts are handled gracefully."""
         # Mock S3 client to return valid SBOM data
@@ -68,7 +68,7 @@ class VulnerabilityScanningTimeoutTests(TestCase):
 
     @patch("sbomify.apps.core.object_store.S3Client")
     @patch("sbomify.apps.vulnerability_scanning.clients.subprocess.run")
-    @patch("sbomify.tasks.dramatiq.get_broker")
+    @patch("sbomify.apps.vulnerability_scanning.tasks.dramatiq.get_broker")
     def test_successful_scan_with_timeout_setting(self, mock_broker, mock_subprocess, mock_s3_client):
         """Test that successful scans use the timeout setting."""
         # Mock S3 client to return valid SBOM data
@@ -107,7 +107,7 @@ class VulnerabilityScanningTimeoutTests(TestCase):
     @override_settings(OSV_SCANNER_TIMEOUT_SECONDS=60)
     @patch("sbomify.apps.core.object_store.S3Client")
     @patch("sbomify.apps.vulnerability_scanning.clients.subprocess.run")
-    @patch("sbomify.tasks.dramatiq.get_broker")
+    @patch("sbomify.apps.vulnerability_scanning.tasks.dramatiq.get_broker")
     def test_custom_timeout_setting(self, mock_broker, mock_subprocess, mock_s3_client):
         """Test that custom timeout settings are respected."""
         # Mock S3 client to return valid SBOM data
